third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-drawImage-video-with-options-expected.txt

Ensure correct behavior of drawImage with ImageBitmaps from video elements along with flipY option.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Checking HTMLVideoElement with imageOrientation: none with premultiplyAlphaOption: default.
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
Checking HTMLVideoElement with imageOrientation: none with premultiplyAlphaOption: none.
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
Checking HTMLVideoElement with imageOrientation: none with premultiplyAlphaOption: invalid.
PASS createImageBitmap with invalid options are rejected
Checking HTMLVideoElement with imageOrientation: flipY with premultiplyAlphaOption: default.
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
Checking HTMLVideoElement with imageOrientation: flipY with premultiplyAlphaOption: none.
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is opaque
PASS This pixel is opaque
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
PASS This pixel is transparent black
Checking HTMLVideoElement with imageOrientation: flipY with premultiplyAlphaOption: invalid.
PASS createImageBitmap with invalid options are rejected
Checking HTMLVideoElement with imageOrientation: invalid with premultiplyAlphaOption: default.
PASS createImageBitmap with invalid options are rejected
Checking HTMLVideoElement with imageOrientation: invalid with premultiplyAlphaOption: none.
PASS createImageBitmap with invalid options are rejected
Checking HTMLVideoElement with imageOrientation: invalid with premultiplyAlphaOption: invalid.
PASS createImageBitmap with invalid options are rejected
PASS successfullyParsed is true

TEST COMPLETE
